split out DisplayStyleSettings class into its own file
