+++ Lib/defconQt/glyphView.py
@@ -29,33 +29,10 @@ from defconQt.addComponentDialog import AddComponentDialog
 from defconQt.addLayerDialog import AddLayerDialog
 from defconQt.anchorItem import AnchorItem
 from defconQt.componentItem import ComponentItem
-from defconQt.genericSettings import GenericSettings
+from defconQt.displayStyleSettings import DisplayStyleSettings
 from defconQt.gotoDialog import GotoDialog
 from defconQt.layerActionsDialog import LayerActionsDialog
 from defconQt.pixmapItem import PixmapItem
-
-class DisplayStyleSettings(GenericSettings):
-    _presets = (
-        ('Default', dict(
-            activeLayerOnTop=True, activeLayerFilled=True,
-            otherLayersFilled=False, activeLayerUseLayerColor=False,
-            otherLayerUseLayerColor=True, drawOtherLayers=True
-        )), ('Layer Fonts', dict(
-            activeLayerOnTop=False, activeLayerFilled=True,
-            otherLayersFilled=True, activeLayerUseLayerColor=True,
-            otherLayerUseLayerColor=True, drawOtherLayers=True
-        ))
-    )
-
-    _items = (
-        ('activeLayerOnTop', ('Active Layer on Top', True)),
-        ('activeLayerFilled', ('Active Layer Filled', True)),
-        ('activeLayerUseLayerColor', ('Active Layer use Custom Color', False)),
-        ('otherLayersFilled', ('Other Layers Filled', False)),
-        ('otherLayerUseLayerColor', ('Other Layers use Custom Color', True)),
-        ('drawOtherLayers', ('Show Other Layers', True))
-    )
-
 
 class MainGfxWindow(QMainWindow):
